requirements.txt
tweepy
transformers
torch
python-dotenv
schedule
aiohttp
